requirements.txt
django
djangorestframework
web3
python3_gearman

